+++ markets/perps-market/contracts/storage/KeeperCosts.sol
@@ -41,7 +41,7 @@ library KeeperCosts {
     function getSettlementKeeperCosts(
         Data storage self,
         uint128 accountId
-    ) internal view returns (uint sUSDCost) {
+    ) internal view returns (uint256 sUSDCost) {
         PerpsMarketFactory.Data storage factory = PerpsMarketFactory.load();
 
         accountId; // unused for now, but will be used to calculate rewards based on account collaterals in the future
@@ -52,14 +52,14 @@ library KeeperCosts {
     function getFlagKeeperCosts(
         Data storage self,
         uint128 accountId
-    ) internal view returns (uint sUSDCost) {
+    ) internal view returns (uint256 sUSDCost) {
         PerpsMarketFactory.Data storage factory = PerpsMarketFactory.load();
 
         PerpsAccount.Data storage account = PerpsAccount.load(accountId);
-        uint numberOfCollateralFeeds = account.activeCollateralTypes.contains(SNX_USD_MARKET_ID)
+        uint256 numberOfCollateralFeeds = account.activeCollateralTypes.contains(SNX_USD_MARKET_ID)
             ? account.activeCollateralTypes.length() - 1
             : account.activeCollateralTypes.length();
-        uint numberOfUpdatedFeeds = numberOfCollateralFeeds +
+        uint256 numberOfUpdatedFeeds = numberOfCollateralFeeds +
             account.openPositionMarketIds.length();
 
         sUSDCost = _processWithRuntime(
@@ -70,7 +70,7 @@ library KeeperCosts {
         );
     }
 
-    function getLiquidateKeeperCosts(Data storage self) internal view returns (uint sUSDCost) {
+    function getLiquidateKeeperCosts(Data storage self) internal view returns (uint256 sUSDCost) {
         PerpsMarketFactory.Data storage factory = PerpsMarketFactory.load();
 
         sUSDCost = _processWithRuntime(self.keeperCostNodeId, factory, 0, KIND_LIQUIDATE);
@@ -81,7 +81,7 @@ library KeeperCosts {
         PerpsMarketFactory.Data storage factory,
         uint256 numberOfUpdatedFeeds,
         uint256 executionKind
-    ) private view returns (uint sUSDCost) {
+    ) private view returns (uint256 sUSDCost) {
         bytes32[] memory runtimeKeys = new bytes32[](4);
         bytes32[] memory runtimeValues = new bytes32[](4);
         runtimeKeys[0] = bytes32("numberOfUpdatedFeeds");
